US30960 Update My Courses
This is a hotfix for 10.8.4 which updates My Courses. The new version of My Courses (v5.5.3) includes a fix for cross-domain requests, which is required when an institution is using a vanity URL. This fix is already in 10.8.5, but will hopefully be removed from there soon in favour of something more resilient and server-side (i.e. generating links that are already on the correct domain, the request's domain).